When opening a line above, don't reindent the current line
For some languages, we can't call `indent-according-to-mode` as there

are multiple possible indentations. Python is a great example. For these

languages, `electric-indent-inhibit` is `t` to prevent emacs changing

indentation without the user requesting it.



In this situation, we give the new line the same indentation characters as the

current line. This ensures we respect the user's tab/space configuration.



Closes #26.