Use read-from-minibuffer for the new file name
`read-file-name` works great for existing file names, but this command
is often used to renaming files to names that aren't currently used.

If a user is using a completion backend like ido or helm,
`read-file-name` will prompt them for an *existing* file name. Instead,
allow them to input arbitrary strings.