Move to beginning of whitespace or end of prompt
Defines crux-smart-line-start to replace back-to-indentation.

back-to-indentation moves the point to the first non-whitespace
character. The new function does that for all major modes, except
ansiterm and eshell, where it moves to end of the prompt. More major
modes can be added.

By fixing this function, crux-move-beginning-of-line is
improved. prelude issue #998 "C-a broken in eshell" is fixed [1].

[1] - https://github.com/bbatsov/prelude/issues/998