Saner functions for modifying cases on region for humans
By default `upcase-region` `downcase-region` and `captialize-region` commands
are disabled in emacs, and for good reason because they are insane. Whenever
you've placed a mark anywhere in the document, inadventently pressing one of the
assigned keybinding will change the cases for all the words in half a
document. These commands resurrect these potentially useful functions in a much
saner way by wrapping them around `use-region-p`, so they are only activated
when `transient-mark-mode` is one.