Optionally move visually in `move-to-mode-line-start'
In some environments users expect crux's functions to respect visual lines. This
can now be configured through the `crux-should-move-visually' configurable
variable. If set to true, then `move-to-mode-line-start' will use take into
account visual lines. Otherwise (the default behavior), it will work as usual by
moving to the logical beginning of the line.

For functions like `crux-move-beginning-of-line', specifically, the user will
face the following situation when setting `crux-should-move-visually; to true:

1. If the user runs this function in the middle of a visual line, it will go to
   the beginning of said visual line.
2. If the user runs this function again, then it will go to the beginning of the
   real line.
3. If the user runs this function again, then it will go to the first non-space
   character of the line.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>