Update crux-transpose-window with other-window
Hello,

The current crux-transpose-window uses next-window and previous-window.
I suggest to rather uses other-window.

The difference between those it that other-window will check for the
no-other-window window parameter.

This parameter is by default, not set to any window. It is at the user
or external packages discretion to set it, with display-buffer-alist or
whatever. So essentially, for a default configuration, this commit
changes nothing.

When the user or a package customizes the window display to set it to
some windows, these window will be skipped by functions such as
other-window, windmove, display-buffer-in-direction.

Essentially one may want to narrow the candidates of commands checking
this parameters to gain in predicability and avoid reusing disposables
window, popups or I don't know how to call them but I think you get the
idea.

I will ramble a little and say this parameter is quite confusing since
you may expect the commands of the ctl-x-4-map to uses it, but no. These
commands generally relies on pop-to-buffer and never check the
no-other-window parameter.

So, maybe too specific for crux ?