Used `derived-mode-p' for functions to also work for derived modes
In my case I recently started using `dirvish', which is an enhanced dired-mode,
and I was surprised that `crux-open-with' didn't work, but I fixed it by using
`derived-mode-p'.

Then, I searched for other occurances of `major-mode' and also replaced

    (equal major-mode '<mode>)

by

    (derived-mode-p '<mode>)

I'm still not sure whether I should also do that in
`crux-cleanup-buffer-or-region', where in checking the membership in
`crux-untabify-sensitive-modes' and `crux-indent-sensitive-modes', derived modes
are note considered. So far I didn't touch that, since those variables are
customizable anyway.